+++ src/main/java/net/geforcemods/securitycraft/entity/camera/CameraController.java
@@ -3,9 +3,9 @@ package net.geforcemods.securitycraft.entity.camera;
 import java.util.function.Consumer;
 
 import net.geforcemods.securitycraft.ClientHandler;
-import net.geforcemods.securitycraft.ConfigHandler;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.IModuleInventory;
+import net.geforcemods.securitycraft.blockentities.SecurityCameraBlockEntity;
 import net.geforcemods.securitycraft.blocks.SecurityCameraBlock;
 import net.geforcemods.securitycraft.misc.KeyBindings;
 import net.geforcemods.securitycraft.misc.LayerToggleHandler;
@@ -41,11 +41,6 @@ import net.neoforged.neoforge.network.PacketDistributor;
 
 @EventBusSubscriber(modid = SecurityCraft.MODID, value = Dist.CLIENT)
 public class CameraController {
-	/**
-	 * @deprecated Don't use directly, use {@link #cameraSpeed()} so the config value gets loaded at the correct time
-	 */
-	@Deprecated
-	private static double cameraSpeed = -1.0D;
 	public static CameraType previousCameraType;
 	public static boolean resetOverlaysAfterDismount = false;
 	private static ClientChunkCache.Storage cameraStorage;
@@ -57,8 +52,8 @@ public class CameraController {
 				//@formatter:off
 				new ViewMovementKeyHandler(mc.options.keyUp, CameraController::moveViewUp),
 				new ViewMovementKeyHandler(mc.options.keyDown, CameraController::moveViewDown),
-				new ViewMovementKeyHandler(mc.options.keyLeft, cam -> moveViewHorizontally(cam, cam.getYRot() - (float) cameraSpeed() * cam.zoomAmount)),
-				new ViewMovementKeyHandler(mc.options.keyRight, cam -> moveViewHorizontally(cam, cam.getYRot() + (float) cameraSpeed() * cam.zoomAmount))
+				new ViewMovementKeyHandler(mc.options.keyLeft, cam -> moveViewHorizontally(cam, cam.getYRot() - getMovementSpeed(cam) * cam.zoomAmount)),
+				new ViewMovementKeyHandler(mc.options.keyRight, cam -> moveViewHorizontally(cam, cam.getYRot() + getMovementSpeed(cam) * cam.zoomAmount))
 				//@formatter:on
 			};
 		}
@@ -142,7 +137,7 @@ public class CameraController {
 	}
 
 	public static void moveViewUp(SecurityCamera cam) {
-		float next = cam.getXRot() - (float) cameraSpeed() * cam.zoomAmount;
+		float next = cam.getXRot() - getMovementSpeed(cam) * cam.zoomAmount;
 
 		if (cam.isCameraDown()) {
 			if (next > 40F)
@@ -153,7 +148,7 @@ public class CameraController {
 	}
 
 	public static void moveViewDown(SecurityCamera cam) {
-		float next = cam.getXRot() + (float) cameraSpeed() * cam.zoomAmount;
+		float next = cam.getXRot() + getMovementSpeed(cam) * cam.zoomAmount;
 
 		if (cam.isCameraDown()) {
 			if (next < 90F)
@@ -235,11 +230,13 @@ public class CameraController {
 		}
 	}
 
-	private static double cameraSpeed() {
-		if (cameraSpeed < 0.0D)
-			cameraSpeed = ConfigHandler.CLIENT.cameraSpeed.get();
+	public static float getMovementSpeed(SecurityCamera cam) {
+		SecurityCameraBlockEntity be = cam.getBlockEntity();
 
-		return cameraSpeed;
+		if (be != null)
+			return (float) be.getMovementSpeed();
+
+		return 0.0F;
 	}
 
 	public static class ViewMovementKeyHandler {