+++ src/main/java/net/geforcemods/securitycraft/entity/camera/SecurityCamera.java
@@ -30,6 +30,7 @@ public class SecurityCamera extends Entity {
 	protected boolean zooming = false;
 	private ChunkTrackingView cameraChunks = null;
 	private boolean hasSentChunks = false;
+	private SecurityCameraBlockEntity be;
 
 	public SecurityCamera(EntityType<SecurityCamera> type, Level level) {
 		super(SCContent.SECURITY_CAMERA_ENTITY.get(), level);
@@ -48,11 +49,13 @@ public class SecurityCamera extends Entity {
 		double y = pos.getY() + 0.5D;
 		double z = pos.getZ() + 0.5D;
 
-		if (cam.isDown())
+		be = cam;
+
+		if (be.isDown())
 			y += 0.25D;
 
 		setPos(x, y, z);
-		setRot(cam.getInitialYRotation(), cam.getInitialXRotation());
+		setRot(be.getInitialYRotation(), be.getInitialXRotation());
 	}
 
 	@Override
@@ -73,7 +76,7 @@ public class SecurityCamera extends Entity {
 	}
 
 	public boolean isCameraDown() {
-		return level().getBlockEntity(blockPosition()) instanceof SecurityCameraBlockEntity cam && cam.isDown();
+		return getBlockEntity() != null && !be.isRemoved() && be.isDown();
 	}
 
 	public void setRotation(float yaw, float pitch) {
@@ -124,8 +127,8 @@ public class SecurityCamera extends Entity {
 	@Deprecated
 	public void discardCamera() {
 		if (!level().isClientSide) {
-			if (level().getBlockEntity(blockPosition()) instanceof SecurityCameraBlockEntity camBe)
-				camBe.stopViewing();
+			if (getBlockEntity() != null && !be.isRemoved())
+				be.stopViewing();
 
 			SectionPos chunkPos = SectionPos.of(blockPosition());
 			int chunkLoadingDistance = cameraChunks instanceof ChunkTrackingView.Positioned cameraChunks ? cameraChunks.viewDistance() : level().getServer().getPlayerList().getViewDistance();
@@ -155,5 +158,16 @@ public class SecurityCamera extends Entity {
 	@Override
 	public boolean isAlwaysTicking() {
 		return true;
+	}
+
+	public SecurityCameraBlockEntity getBlockEntity() {
+		if (be == null) {
+			if (level().getBlockEntity(blockPosition()) instanceof SecurityCameraBlockEntity camera)
+				be = camera;
+			else
+				SecurityCraft.LOGGER.warn("No security camera block entity was found at {}. Try breaking and replacing the block!", blockPosition());
+		}
+
+		return be;
 	}
 }