+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for v1.9.10 of SecurityCraft--------------------------
 
+- Change: The cameraSpeed client side config setting has been moved to be a per-block option, accessible with the Universal Block Modifier
 - Fix: Trying to place a Panic Button on top of powdered snow crashes the game
 
 --------------------------Changelog for v1.9.9-beta2 of SecurityCraft--------------------------